feat: Add session review widgets and enhance live session UI
