+++ lib/features/history/session_export.dart
@@ -1,23 +1,21 @@
 // =============================================================================
-// Session Export — Raven selection table and ZIP bundle for sharing
+// Session Export
 // =============================================================================
 //
 // Generates export artifacts for a live session:
 //
 //   • **Raven selection table** (.txt): Tab-delimited annotation file
-//     compatible with Raven Pro / Raven Lite (Cornell Lab). Each detection
-//     becomes a row with begin/end time offsets, frequency bounds, species
-//     names, and confidence.
+//     compatible with Raven Pro / Raven Lite.
 //
-//   • **ZIP bundle** (.zip): Archives the full WAV recording together with
-//     the selection table for convenient sharing.
+//   • **CSV Export** (.csv): Standard comma-separated values.
 //
-// The selection table follows the standard Raven format with BirdNET-specific
-// columns (Common Name, Scientific Name, Confidence) appended after the core
-// columns (Selection, View, Channel, Begin Time, End Time, Low Freq, High
-// Freq).
+//   • **JSON Export** (.json): Machine-readable JSON structured data.
+//
+//   • **ZIP bundle** (.zip): Optionally archives the full WAV/FLAC recording
+//     together with the export document for convenient sharing.
 // =============================================================================
 
+import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
@@ -29,11 +27,7 @@ import '../live/live_session.dart';
 /// Upper frequency bound for Raven annotations (Nyquist of 32 kHz).
 const int _highFreqHz = 16000;
 
-/// Generates a Raven Pro–compatible selection table from session detections.
-///
-/// Returns the table as a UTF-8 string with tab-separated columns:
-///   Selection, View, Channel, Begin Time (s), End Time (s),
-///   Low Freq (Hz), High Freq (Hz), Common Name, Scientific Name, Confidence
+/// Generates a Raven Pro-compatible selection table from session detections.
 String buildRavenSelectionTable(LiveSession session) {
   final buf = StringBuffer();
 
@@ -50,7 +44,6 @@ String buildRavenSelectionTable(LiveSession session) {
   for (var i = 0; i < session.detections.length; i++) {
     final d = session.detections[i];
 
-    // Offset from recording start in seconds.
     final beginSec =
         d.timestamp.difference(session.startTime).inMilliseconds / 1000.0;
     final endSec = beginSec + windowSeconds;
@@ -72,39 +65,126 @@ String buildRavenSelectionTable(LiveSession session) {
   return buf.toString();
 }
 
-/// Creates a ZIP archive containing the session audio (WAV or FLAC) and a
-/// Raven selection table, and writes it to a temporary file.
+/// Generates a standard CSV representation of session detections.
+String buildCsvExport(LiveSession session) {
+  final buf = StringBuffer();
+
+  // Header row.
+  buf.writeln(
+      'Timestamp,Begin Time (s),Common Name,Scientific Name,Confidence');
+
+  for (final d in session.detections) {
+    final beginSec =
+        d.timestamp.difference(session.startTime).inMilliseconds / 1000.0;
+
+    // Simple CSV escaping for species names
+    final commonName =
+        d.commonName.contains(',') ? '"${d.commonName}"' : d.commonName;
+    final sciName = d.scientificName.contains(',')
+        ? '"${d.scientificName}"'
+        : d.scientificName;
+
+    buf.writeln(
+      '${d.timestamp.toIso8601String()},'
+      '${beginSec.toStringAsFixed(3)},'
+      '$commonName,'
+      '$sciName,'
+      '${d.confidence.toStringAsFixed(4)}',
+    );
+  }
+
+  return buf.toString();
+}
+
+/// Generates a JSON representation of the session and its detections.
+String buildJsonExport(LiveSession session) {
+  final map = {
+    'session': session.displayName,
+    'startTime': session.startTime.toIso8601String(),
+    'endTime': session.endTime?.toIso8601String(),
+    'recordingPath': session.recordingPath,
+    'settings': {
+      'windowDuration': session.settings.windowDuration,
+      'confidenceThreshold': session.settings.confidenceThreshold,
+      'inferenceRate': session.settings.inferenceRate,
+      'speciesFilterMode': session.settings.speciesFilterMode,
+    },
+    'detections': session.detections.map((d) {
+      final beginSec =
+          d.timestamp.difference(session.startTime).inMilliseconds / 1000.0;
+      return {
+        'timestamp': d.timestamp.toIso8601String(),
+        'beginTimeSec': num.parse(beginSec.toStringAsFixed(3)),
+        'commonName': d.commonName,
+        'scientificName': d.scientificName,
+        'confidence': num.parse(d.confidence.toStringAsFixed(4)),
+      };
+    }).toList(),
+  };
+
+  return const JsonEncoder.withIndent('  ').convert(map);
+}
+
+/// Creates an export bundle containing the session data and optionally audio.
 ///
-/// Returns the path to the ZIP file, or `null` if the recording does not exist.
-Future<String?> buildSessionZip(LiveSession session) async {
-  final audioPath = session.recordingPath;
-  if (audioPath == null || !File(audioPath).existsSync()) return null;
-
-  final archive = Archive();
-
-  // Use the session display name for all exported filenames.
+/// If [includeAudio] is true and audio exists, returns a path to a .zip file.
+/// If [includeAudio] is false or no audio exists, returns a path to the raw
+/// text/json file.
+Future<String?> buildSessionExport(
+  LiveSession session, {
+  required String format,
+  required bool includeAudio,
+}) async {
   final baseName = session.displayName;
-  final audioExt = p.extension(audioPath); // .wav or .flac
+  final audioPath = session.recordingPath;
+  final hasAudio = audioPath != null && File(audioPath).existsSync();
 
-  // Add the audio file (WAV or FLAC).
-  final audioBytes = await File(audioPath).readAsBytes();
-  archive.addFile(
-    ArchiveFile('$baseName$audioExt', audioBytes.length, audioBytes),
-  );
+  String fileContent;
+  String extension;
 
-  // Add the Raven selection table.
-  final table = buildRavenSelectionTable(session);
-  final tableBytes = Uint8List.fromList(table.codeUnits);
-  archive.addFile(
-    ArchiveFile('$baseName.selections.txt', tableBytes.length, tableBytes),
-  );
+  switch (format) {
+    case 'csv':
+      fileContent = buildCsvExport(session);
+      extension = '.csv';
+      break;
+    case 'json':
+      fileContent = buildJsonExport(session);
+      extension = '.json';
+      break;
+    case 'raven':
+    default:
+      fileContent = buildRavenSelectionTable(session);
+      extension = '.selections.txt';
+      break;
+  }
 
-  // Encode and write to a temp file alongside the audio.
-  final zipBytes = ZipEncoder().encode(archive);
+  final bytes = Uint8List.fromList(utf8.encode(fileContent));
 
-  final zipName = '$baseName.zip';
-  final zipPath = p.join(p.dirname(audioPath), zipName);
-  await File(zipPath).writeAsBytes(zipBytes);
+  if (includeAudio && hasAudio) {
+    final archive = Archive();
+    final audioExt = p.extension(audioPath);
+    final audioBytes = await File(audioPath).readAsBytes();
 
-  return zipPath;
+    archive.addFile(
+      ArchiveFile('$baseName$audioExt', audioBytes.length, audioBytes),
+    );
+    archive.addFile(
+      ArchiveFile('$baseName$extension', bytes.length, bytes),
+    );
+
+    final zipBytes = ZipEncoder().encode(archive);
+
+    final zipName = '$baseName.zip';
+    final zipPath = p.join(p.dirname(audioPath), zipName);
+    await File(zipPath).writeAsBytes(zipBytes!);
+
+    return zipPath;
+  } else {
+    // If no audio or user opted out of including audio, just write and share the doc file.
+    final dir = hasAudio ? p.dirname(audioPath) : Directory.systemTemp.path;
+    final filePath = p.join(dir, '$baseName$extension');
+    await File(filePath).writeAsBytes(bytes);
+
+    return filePath;
+  }
 }
